requirements.txt
web.py==0.62
pyrebase4==4.5.0
gunicorn==20.1.0